+++ cmd/vega/genesis/genesis.go
@@ -15,8 +15,6 @@ type Cmd struct {
 	// Subcommands
 	Generate       generateCmd       `command:"generate" description:"Generates the genesis file"`
 	Update         updateCmd         `command:"update" description:"Update the genesis file with the app_state, useful if the genesis generation is not done using \"vega genesis generate\""`
-	Sign           signCmd           `command:"sign" description:"Sign a subset of the network parameters"`
-	Verify         verifyCmd         `command:"verify" description:"Verify the signature of the network parameter against local genesis file"`
 	LoadCheckpoint loadCheckpointCmd `command:"load_checkpoint" description:"Load the given checkpoint file in the genesis file"`
 }
 
@@ -25,12 +23,6 @@ var genesisCmd Cmd
 func Genesis(ctx context.Context, parser *flags.Parser) error {
 	genesisCmd = Cmd{
 		Generate: generateCmd{
-			TmHome: "$HOME/.tendermint",
-		},
-		Sign: signCmd{
-			TmHome: "$HOME/.tendermint",
-		},
-		Verify: verifyCmd{
 			TmHome: "$HOME/.tendermint",
 		},
 		Update: updateCmd{