+++ liquidity/snapshot_test.go
@@ -95,7 +95,7 @@ func TestSnapshotRoundTrip(t *testing.T) {
 		"partiesLiquidityOrders:market-id": "0254d8b74441ca3bac8f9b141408502d9b1f297e8ef1054d45775566677a8072",
 		"partiesOrders:market-id":          "f9cb31b1c4c8df91f6a348d43978c302c8887336107c265259bc74fdddf00e19",
 		"pendingProvisions:market-id":      "6cc4d407a2ea45e37e27993eb6f94134b3f906d080777d94bf99551aa82dc461",
-		"provisions:market-id":             "20b1f6850f633a7a304618dd9a85e6433af194757b8294e2566bd4bf5fb8423c",
+		"provisions:market-id":             "db6161ad5d878863d2a89269c5153642a5e1f86f65cbae5cc4d037d33964e238",
 		"liquiditySupplied:market-id":      "3276bba2a77778ba710ec29e3a6e59212452dbda69eaac8f9160930d1270da1d",
 	}
 
@@ -133,7 +133,7 @@ func TestSnapshotRoundTrip(t *testing.T) {
 		"partiesLiquidityOrders:market-id": "c92b7c102539b6c59b2b3d4be17cabbdf4b385fa85007f69e6092bbc68c186b6",
 		"partiesOrders:market-id":          "f9cb31b1c4c8df91f6a348d43978c302c8887336107c265259bc74fdddf00e19",
 		"pendingProvisions:market-id":      "627ef55af7f36bea0d09b0081b85d66531a01df060d8e9447e17049a4e152b12",
-		"provisions:market-id":             "7f793b051cd4dc61c94f3193da156c09eb38ac206b0428eccca26a2d555b303b",
+		"provisions:market-id":             "1f8dd286b9d1cbfeaec6dfe67881f8959885acc2fe561816017764f2b5e9d149",
 		"liquiditySupplied:market-id":      "3276bba2a77778ba710ec29e3a6e59212452dbda69eaac8f9160930d1270da1d",
 	}