+++ validators/validator_set.go
@@ -140,11 +140,7 @@ func (t *Topology) RecalcValidatorSet(ctx context.Context, epochSeq string, dele
 		}
 	}
 
-	// do this only if we are a validator, not need otherwise
-	if t.IsValidator() {
-		t.signatures.EmitPromotionsSignatures(
-			ctx, t.currentTime, currentState, newState)
-	}
+	t.signatures.EmitPromotionsSignatures(ctx, t.currentTime, t.epochSeq, currentState, newState)
 
 	// prepare and send the events
 	evts := make([]events.Event, 0, len(currentState))